src/test/java/wniemiec/mobilex/ama/models/behavior/SpreadElementTest.java
package wniemiec.mobilex.ama.models.behavior;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;


class SpreadElementTest {
    
    //-------------------------------------------------------------------------
    //		Attributes
    //-------------------------------------------------------------------------
    private SpreadElement spreadElement;
    private Expression argument;


    //-------------------------------------------------------------------------
    //		Test hooks
    //-------------------------------------------------------------------------
    @BeforeEach
    void setUp() {
        spreadElement = null;
        argument = null;
    }


    //-------------------------------------------------------------------------
    //		Tests
    //-------------------------------------------------------------------------
    @Test
    void testToCodeWithArgument() {
        withArgument(new Identifier("foo"));
        buildSpreadElement();
        assertToCodeIs("...foo");
    }

    
    //-------------------------------------------------------------------------
    //		Methods
    //-------------------------------------------------------------------------
    private void withArgument(Expression expression) {
        argument = expression;
    }

    private void buildSpreadElement() {
        spreadElement = new SpreadElement(argument);
    }

    private void assertToCodeIs(String expectedCode) {
        assertHasSameLine(expectedCode, spreadElement.toCode());
    }

    private void assertHasSameLine(String expected, String obtained) {
        Assertions.assertEquals(
            removeWhiteSpaces(expected),
            removeWhiteSpaces(obtained)
        );
    }

    private String removeWhiteSpaces(String text) {
        return text.replaceAll("[\\s\\t]+", "");
    }
}
